chore(fileStructure): organise files/folder to follow a defined file structure
- group (files) by features or routes

[Finishes #163053732]
