+++ .eslintignore
@@ -1,2 +1,1 @@
-__tests__/**
 **/utils/index.js
