+++ src/components/menu/Menu.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import propTypes from 'prop-types';
-import Button from './Button';
+import Button from '../presentation/Button';
 
 const Menu = (props) => {
   const {
