+++ src/components/menu/Menu.test.js
@@ -2,18 +2,20 @@ import React from 'react';
 import { render } from 'react-testing-library';
 import 'react-testing-library/cleanup-after-each';
 import 'jest-dom/extend-expect';
-import Menu from '../../src/components/presentation/Menu';
+import Menu from './Menu';
 
 describe('<Menu />', () => {
   const props = {
     name: 'rice',
     imgurl: 'https://pix.io/rice.jpg',
-    price: 300
+    price: 300,
   };
 
   let renderResult = {};
 
-  beforeEach(() => (renderResult = render(<Menu menu={props} />)));
+  beforeEach(() => {
+    renderResult = render(<Menu menu={props} />);
+  });
 
   test('renders', () => {
     const { getByTestId } = render(<Menu menu={{}} />);
@@ -26,10 +28,7 @@ describe('<Menu />', () => {
 
     expect(container.querySelector('.menu__name')).toHaveTextContent(name);
     expect(container.querySelector('.menu__price')).toHaveTextContent(price);
-    expect(container.querySelector('.menu__img>img')).toHaveAttribute(
-      'src',
-      imgurl
-    );
+    expect(container.querySelector('.menu__img>img')).toHaveAttribute('src', imgurl);
     expect(getByAltText(name)).toBeInTheDocument();
   });
