+++ src/index.js
@@ -5,7 +5,7 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import './assets/sass/style.scss';
 import reducers from './reducers';
-import AppRouter from './routes';
+import AppRouter from './components/route';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose; // eslint-disable-line
 const store = createStore(reducers, composeEnhancers(applyMiddleware(thunk)));
@@ -16,4 +16,4 @@ const appProvider = (
   </Provider>
 );
 
-ReactDOM.render(appProvider, document.getElementById('app')); // eslint-disable-line
+ReactDOM.render(appProvider, document.getElementById('app'));
